moved logger and validate to foundation
