+++ app/services/sales-api/handlers/v1/productgrp/filter.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 
 	"github.com/ardanlabs/service/business/core/product"
-	"github.com/ardanlabs/service/business/sys/validate"
+	"github.com/ardanlabs/service/foundation/validate"
 	"github.com/google/uuid"
 )
