+++ app/services/sales-api/handlers/v1/productgrp/order.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/ardanlabs/service/business/core/product"
 	"github.com/ardanlabs/service/business/data/order"
-	"github.com/ardanlabs/service/business/sys/validate"
+	"github.com/ardanlabs/service/foundation/validate"
 )
 
 var orderByFields = map[string]struct{}{
