+++ business/core/product/filter.go
@@ -3,7 +3,7 @@ package product
 import (
 	"fmt"
 
-	"github.com/ardanlabs/service/business/sys/validate"
+	"github.com/ardanlabs/service/foundation/validate"
 	"github.com/google/uuid"
 )
