+++ foundation/logger/model.go
